fix swerve

+++ src/main/java/frc/robot/Constants.java
@@ -282,7 +282,7 @@ public final class Constants {
 
         public static final double OUTTAKE_TIME = 0.2;
 
-        public static final double INTAKE_MOTOR = 50;
+        public static final int INTAKE_MOTOR = 50;
         
         public static final Color CUBE_COLOR = new Color(0.21, 0.33, 0.46);
         public static final Color CONE_COLOR = new Color(0.37, 0.57, 0.00); // grayson cone: new Color(141, 93, 20);
